Add new SerializedSize trait  (#781)

+++ src/chain-libs/chain-impl-mockchain/src/header/cstruct.rs
@@ -7,7 +7,7 @@ use std::mem::size_of;
 // Offset and Basic Types
 // ************************************************************************
 
-pub(super) type Version = u16;
+pub(super) type Version = u8;
 pub(super) type ContentSize = u32;
 pub(super) type DateEpoch = u32;
 pub(super) type DateSlotid = u32;
